+++ packages/transaction-controller/jest.config.js
@@ -17,10 +17,10 @@ module.exports = merge(baseConfig, {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 83.77,
-      functions: 92.56,
-      lines: 95.11,
-      statements: 95.22,
+      branches: 84.02,
+      functions: 93.33,
+      lines: 95.61,
+      statements: 95.71,
     },
   },