+++ packages/transaction-controller/src/TransactionController.test.ts
@@ -121,7 +121,12 @@ jest.mock('@metamask/eth-query', () =>
       },
       getBlockByHash: (_blockHash: any, callback: any) => {
         const blocks: any = [
-          { hash: '1337', number: '0x1', baseFeePerGas: '0x14' },
+          {
+            hash: '1337',
+            number: '0x1',
+            baseFeePerGas: '0x14',
+            timestamp: '628dc0c8',
+          },
           { hash: '1338', number: '0x2' },
         ];
         const block: any = blocks.find(
@@ -383,7 +388,9 @@ const MOCK_CUSTOM_NETWORK: MockNetwork = {
 };
 
 const ACCOUNT_MOCK = '0x6bf137f335ea1b8f193b8f6ea92561a60d23a207';
+const ACCOUNT_2_MOCK = '0x08f137f335ea1b8f193b8f6ea92561a60d23a211';
 const NONCE_MOCK = 12;
+const ACTION_ID_MOCK = '123456';
 
 const TRANSACTION_META_MOCK = {
   status: TransactionStatus.confirmed,
@@ -693,12 +700,218 @@ describe('TransactionController', () => {
     });
   });
 
+  describe('with actionId', () => {
+    it('adds single unapproved transaction when called twice with same actionId', async () => {
+      const controller = newController();
+
+      const mockOrigin = 'origin';
+
+      await controller.addTransaction(
+        {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        },
+        {
+          origin: mockOrigin,
+          actionId: ACTION_ID_MOCK,
+        },
+      );
+
+      const firstTransactionCount = controller.state.transactions.length;
+
+      await controller.addTransaction(
+        {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        },
+        {
+          origin: mockOrigin,
+          actionId: ACTION_ID_MOCK,
+        },
+      );
+      const secondTransactionCount = controller.state.transactions.length;
+
+      expect(firstTransactionCount).toStrictEqual(secondTransactionCount);
+      expect(delayMessengerMock.call).toHaveBeenCalledTimes(1);
+      expect(delayMessengerMock.call).toHaveBeenCalledWith(
+        'ApprovalController:addRequest',
+        {
+          id: expect.any(String),
+          origin: mockOrigin,
+          type: 'transaction',
+          requestData: { txId: expect.any(String) },
+          expectsResult: true,
+        },
+        true,
+      );
+    });
+
+    it('adds multiple transactions with same actionId and ensures second transaction result does not resolves before the first transaction result', async () => {
+      const controller = newController();
+
+      const mockOrigin = 'origin';
+      let firstTransactionCompleted = false;
+      let secondTransactionCompleted = false;
+
+      const { result: firstResult } = await controller.addTransaction(
+        {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        },
+        {
+          origin: mockOrigin,
+          actionId: ACTION_ID_MOCK,
+        },
+      );
+
+      firstResult
+        .then(() => {
+          firstTransactionCompleted = true;
+        })
+        .catch(() => undefined);
+
+      const { result: secondResult } = await controller.addTransaction(
+        {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_MOCK,
+        },
+        {
+          origin: mockOrigin,
+          actionId: ACTION_ID_MOCK,
+        },
+      );
+      secondResult
+        .then(() => {
+          secondTransactionCompleted = true;
+        })
+        .catch(() => undefined);
+
+      await wait(0);
+
+      expect(firstTransactionCompleted).toBe(false);
+      expect(secondTransactionCompleted).toBe(false);
+
+      approveTransaction();
+      await firstResult;
+      await secondResult;
+
+      expect(firstTransactionCompleted).toBe(true);
+      expect(secondTransactionCompleted).toBe(true);
+    });
+
+    it.each([
+      [
+        'does not add duplicate transaction if actionId already used',
+        ACTION_ID_MOCK,
+        ACTION_ID_MOCK,
+        1,
+      ],
+      [
+        'adds additional transaction if actionId not used',
+        ACTION_ID_MOCK,
+        '00000',
+        2,
+      ],
+    ])(
+      '%s',
+      async (_, firstActionId, secondActionId, expectedTransactionCount) => {
+        const controller = newController();
+        const expectedRequestApprovalCalledTimes = expectedTransactionCount;
+
+        const mockOrigin = 'origin';
+
+        await controller.addTransaction(
+          {
+            from: ACCOUNT_MOCK,
+            to: ACCOUNT_MOCK,
+          },
+          {
+            origin: mockOrigin,
+            actionId: firstActionId,
+          },
+        );
+
+        await controller.addTransaction(
+          {
+            from: ACCOUNT_MOCK,
+            to: ACCOUNT_MOCK,
+          },
+          {
+            origin: mockOrigin,
+            actionId: secondActionId,
+          },
+        );
+        const { transactions } = controller.state;
+
+        expect(transactions).toHaveLength(expectedTransactionCount);
+        expect(delayMessengerMock.call).toHaveBeenCalledTimes(
+          expectedRequestApprovalCalledTimes,
+        );
+      },
+    );
+
+    it.each([
+      [
+        'adds single transaction when speed up called twice with the same actionId',
+        ACTION_ID_MOCK,
+        2,
+        1,
+      ],
+      [
+        'adds multiple transactions when speed up called with non-existent actionId',
+        '00000',
+        3,
+        2,
+      ],
+    ])(
+      '%s',
+      async (
+        _,
+        actionId,
+        expectedTransactionCount,
+        expectedSignCalledTimes,
+      ) => {
+        const controller = newController();
+        const signSpy = jest.spyOn(controller, 'sign' as any);
+
+        const { transactionMeta } = await controller.addTransaction({
+          from: ACCOUNT_MOCK,
+          gas: '0x0',
+          gasPrice: '0x50fd51da',
+          to: ACCOUNT_MOCK,
+          value: '0x0',
+        });
+        await controller.speedUpTransaction(transactionMeta.id, undefined, {
+          actionId: ACTION_ID_MOCK,
+        });
+
+        await controller.speedUpTransaction(transactionMeta.id, undefined, {
+          actionId,
+        });
+
+        const { transactions } = controller.state;
+        expect(transactions).toHaveLength(expectedTransactionCount);
+        expect(signSpy).toHaveBeenCalledTimes(expectedSignCalledTimes);
+      },
+    );
+  });
+
   describe('addTransaction', () => {
     it('adds unapproved transaction to state', async () => {
       const controller = newController();
 
       const mockDeviceConfirmedOn = WalletDevice.OTHER;
       const mockOrigin = 'origin';
+      const mockSecurityAlertResponse = {
+        resultType: 'Malicious',
+        reason: 'blur_farming',
+        description:
+          'A SetApprovalForAll request was made on {contract}. We found the operator {operator} to be malicious',
+        args: {
+          contract: '0xa7206d878c5c3871826dfdb42191c49b1d11f466',
+          operator: '0x92a3b9773b1763efa556f55ccbeb20441962d9b2',
+        },
+      };
 
       await controller.addTransaction(
         {
@@ -708,6 +921,7 @@ describe('TransactionController', () => {
         {
           deviceConfirmedOn: mockDeviceConfirmedOn,
           origin: mockOrigin,
+          securityAlertResponse: mockSecurityAlertResponse,
         },
       );
 
@@ -723,6 +937,9 @@ describe('TransactionController', () => {
       expect(controller.state.transactions[0].origin).toBe(mockOrigin);
       expect(controller.state.transactions[0].status).toBe(
         TransactionStatus.unapproved,
+      );
+      expect(controller.state.transactions[0].securityAlertResponse).toBe(
+        mockSecurityAlertResponse,
       );
     });
 
@@ -947,10 +1164,12 @@ describe('TransactionController', () => {
 
         await result;
 
-        const { transaction, status } = controller.state.transactions[0];
+        const { transaction, status, submittedTime } =
+          controller.state.transactions[0];
         expect(transaction.from).toBe(ACCOUNT_MOCK);
         expect(transaction.nonce).toBe(`0x${NONCE_MOCK.toString(16)}`);
         expect(status).toBe(TransactionStatus.submitted);
+        expect(submittedTime).toBeLessThanOrEqual(Date.now());
       });
 
       it('reports success to approval acceptor', async () => {
@@ -1049,10 +1268,7 @@ describe('TransactionController', () => {
           (
             delayMessengerMock.call as jest.MockedFunction<any>
           ).mockImplementationOnce(() => {
-            controller.state.transactions[0].status =
-              TransactionStatus.confirmed;
-
-            throw new Error('TestError');
+            throw new Error('Unknown problem');
           });
 
           const { result } = await controller.addTransaction({
@@ -1093,7 +1309,7 @@ describe('TransactionController', () => {
             delayMessengerMock.call as jest.MockedFunction<any>
           ).mockImplementationOnce(() => {
             controller.state.transactions = [];
-            throw new Error('TestError');
+            throw new Error('Unknown problem');
           });
 
           const { result } = await controller.addTransaction({
@@ -1268,6 +1484,7 @@ describe('TransactionController', () => {
       const transactionMeta = controller.state.transactions[0];
       expect(transactionMeta.verifiedOnBlockchain).toBe(true);
       expect(transactionMeta.transaction.gasUsed).toBe('0x5208');
+      expect(transactionMeta.blockTimestamp).toBe('628dc0c8');
       expect(transactionMeta.baseFeePerGas).toBe('0x14');
       expect(transactionMeta.txReceipt?.transactionIndex).toBe(1337);
     });
@@ -1366,11 +1583,34 @@ describe('TransactionController', () => {
         value: '0x0',
       });
 
-      await controller.stopTransaction(transactionMeta.id);
+      const finishedPromise = waitForTransactionFinished(controller);
+
+      await controller.stopTransaction(transactionMeta.id, undefined, {
+        estimatedBaseFee: '0x123',
+      });
+
+      const { estimatedBaseFee } = await finishedPromise;
 
       approveTransaction();
 
+      const { transactions } = controller.state;
       await expect(result).rejects.toThrow('User cancelled the transaction');
+      expect(estimatedBaseFee).toBe('0x123');
+      expect(transactions[0].status).toStrictEqual(TransactionStatus.cancelled);
+    });
+
+    it('rejects unknown transaction', async () => {
+      const controller = newController({
+        network: MOCK_LINEA_GOERLI_NETWORK,
+      });
+
+      await controller.stopTransaction('transactionIdMock', {
+        gasPrice: '0x1',
+      });
+
+      const signSpy = jest.spyOn(controller, 'sign' as any);
+
+      expect(signSpy).toHaveBeenCalledTimes(0);
     });
 
     it('throws if no sign method', async () => {
@@ -1407,6 +1647,28 @@ describe('TransactionController', () => {
       );
     });
 
+    it('creates additional transaction specifying the gasPrice', async () => {
+      const controller = newController({
+        network: MOCK_LINEA_MAINNET_NETWORK,
+      });
+
+      const { transactionMeta } = await controller.addTransaction({
+        from: ACCOUNT_MOCK,
+        gas: '0x0',
+        gasPrice: '0x50fd51da',
+        to: ACCOUNT_MOCK,
+        value: '0x0',
+      });
+
+      await controller.speedUpTransaction(transactionMeta.id, {
+        gasPrice: '0x62DEF4DA',
+      });
+
+      const { transactions } = controller.state;
+      expect(transactions).toHaveLength(2);
+      expect(transactions[1].transaction.gasPrice).toBe('0x62DEF4DA');
+    });
+
     it('uses the same nonce', async () => {
       const controller = newController({ approve: true });
 
@@ -1419,7 +1681,9 @@ describe('TransactionController', () => {
       });
 
       await result;
-      await controller.speedUpTransaction(transactionMeta.id);
+      await controller.speedUpTransaction(transactionMeta.id, undefined, {
+        estimatedBaseFee: '0x123',
+      });
 
       const { transactions } = controller.state;
       expect(getNonceLockSpy).toHaveBeenCalledTimes(1);
@@ -1428,6 +1692,7 @@ describe('TransactionController', () => {
       expect(transactions[0].transaction.nonce).toStrictEqual(
         transactions[1].transaction.nonce,
       );
+      expect(transactions[1].estimatedBaseFee).toBe('0x123');
     });
 
     it('allows transaction count to exceed txHistorylimit', async () => {
@@ -1503,6 +1768,158 @@ describe('TransactionController', () => {
       controller.initApprovals();
 
       expect(delayMessengerMock.call).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('confirmExternalTransaction', () => {
+    it('adds external transaction to the state as confirmed', async () => {
+      const controller = newController();
+
+      const externalTransactionToConfirm = {
+        from: ACCOUNT_MOCK,
+        to: ACCOUNT_2_MOCK,
+        id: '1',
+        networkID: '1',
+        chainId: toHex(1),
+        status: TransactionStatus.confirmed,
+        transaction: {
+          gasUsed: undefined,
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_2_MOCK,
+        },
+      } as any;
+      const externalTransactionReceipt = {
+        gasUsed: '0x5208',
+      };
+      const externalBaseFeePerGas = '0x14';
+
+      await controller.confirmExternalTransaction(
+        externalTransactionToConfirm,
+        externalTransactionReceipt,
+        externalBaseFeePerGas,
+      );
+
+      expect(controller.state.transactions[0].status).toBe(
+        TransactionStatus.confirmed,
+      );
+      expect(controller.state.transactions[0].baseFeePerGas).toBe(
+        externalBaseFeePerGas,
+      );
+      expect(controller.state.transactions[0]?.txReceipt?.gasUsed).toBe(
+        externalTransactionReceipt.gasUsed,
+      );
+    });
+
+    it('marks the same nonce local transactions statuses as dropped and defines replacedBy properties', async () => {
+      const controller = newController();
+      const externalTransactionId = '1';
+      const externalTransactionHash = '0x1';
+      const externalTransactionToConfirm = {
+        from: ACCOUNT_MOCK,
+        to: ACCOUNT_2_MOCK,
+        hash: externalTransactionHash,
+        id: externalTransactionId,
+        networkID: '5',
+        chainId: toHex(5),
+        status: TransactionStatus.confirmed,
+        transaction: {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_2_MOCK,
+          nonce: NONCE_MOCK,
+        },
+      } as any;
+      const externalTransactionReceipt = {
+        gasUsed: '0x5208',
+      };
+      const externalBaseFeePerGas = '0x14';
+
+      // Submitted local unapproved transaction
+      const localTransactionIdWithSameNonce = '9';
+      controller.state.transactions.push({
+        from: ACCOUNT_MOCK,
+        to: ACCOUNT_2_MOCK,
+        id: localTransactionIdWithSameNonce,
+        networkID: '5',
+        chainId: toHex(5),
+        status: TransactionStatus.unapproved,
+        transaction: {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_2_MOCK,
+          nonce: NONCE_MOCK,
+        },
+      } as any);
+
+      await controller.confirmExternalTransaction(
+        externalTransactionToConfirm,
+        externalTransactionReceipt,
+        externalBaseFeePerGas,
+      );
+
+      const droppedTx = controller.state.transactions.find(
+        (transaction) => transaction.id === localTransactionIdWithSameNonce,
+      );
+
+      expect(droppedTx?.status).toBe(TransactionStatus.dropped);
+
+      expect(droppedTx?.replacedById).toBe(externalTransactionId);
+
+      expect(droppedTx?.replacedBy).toBe(externalTransactionHash);
+    });
+
+    it('doesnt mark transaction as dropped if same nonce local transaction status is failed', async () => {
+      const controller = newController();
+      const externalTransactionId = '1';
+      const externalTransactionHash = '0x1';
+      const externalTransactionToConfirm = {
+        from: ACCOUNT_MOCK,
+        to: ACCOUNT_2_MOCK,
+        hash: externalTransactionHash,
+        id: externalTransactionId,
+        networkID: '5',
+        chainId: toHex(5),
+        status: TransactionStatus.confirmed,
+        transaction: {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_2_MOCK,
+          nonce: NONCE_MOCK,
+        },
+      } as any;
+      const externalTransactionReceipt = {
+        gasUsed: '0x5208',
+      };
+      const externalBaseFeePerGas = '0x14';
+
+      // Off-chain failed local transaction
+      const localTransactionIdWithSameNonce = '9';
+      controller.state.transactions.push({
+        from: ACCOUNT_MOCK,
+        to: ACCOUNT_2_MOCK,
+        id: localTransactionIdWithSameNonce,
+        networkID: '5',
+        chainId: toHex(5),
+        status: TransactionStatus.failed,
+        transaction: {
+          from: ACCOUNT_MOCK,
+          to: ACCOUNT_2_MOCK,
+          nonce: NONCE_MOCK,
+        },
+      } as any);
+
+      await controller.confirmExternalTransaction(
+        externalTransactionToConfirm,
+        externalTransactionReceipt,
+        externalBaseFeePerGas,
+      );
+
+      const failedTx = controller.state.transactions.find(
+        (transaction) => transaction.id === localTransactionIdWithSameNonce,
+      );
+
+      expect(failedTx?.status).toBe(TransactionStatus.failed);
+
+      expect(failedTx?.replacedById).toBe(externalTransactionId);
+
+      expect(failedTx?.replacedBy).toBe(externalTransactionHash);
     });
   });